Use shortcuts for input, not numbers
Previously, for a given grid we could set a set of shortcuts and a
corresponding number. This doesn't work for grids with more than 10 elements
and also it is hard to associate numbers with objects. Therefore, now we can
only set shortcuts and use them.

In order to count objects, just use numbers as shortcuts.